requirements.txt
# Core dependencies
duckdb>=1.1.3
pyarrow>=18.1.0
pandas>=2.2.3
openpyxl>=3.1.5
pyyaml>=6.0.2
pydantic>=2.5.0
tqdm>=4.66.0

# Development dependencies (optional)
pytest>=7.4.4
pylint>=3.0.3
